dependencymapping/requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
psycopg2-binary
django-grappelli==2.11.1
django-rest-swagger==2.2.0
pytz==2018.3
pytest==3.8.0
pytest-watch==4.2.0
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0
